+++ apps/recnet/src/app/feeds/CutoffDatePicker.tsx
@@ -98,8 +98,9 @@ export function CutoffDatePicker(props: CutoffDatePickerProps) {
         open={isOpen}
         onOpenChange={(open) => {
           setIsOpen(open);
-          // reset view on open
+          // reset view on open and reset to today
           if (open) {
+            navigation.setToday();
             setView("default");
           }
         }}
@@ -109,7 +110,11 @@ export function CutoffDatePicker(props: CutoffDatePickerProps) {
             More
           </Button>
         </Popover.Trigger>
-        <Popover.Content className="overflow-hidden">
+        <Popover.Content
+          className="overflow-hidden"
+          side="right"
+          alignOffset={-50}
+        >
           <Flex className="justify-center items-center w-full p-2 mb-2">
             <Button
               variant="ghost"